+++ sys/include/machine/signal.h
@@ -10,6 +10,8 @@
 #define _MACHINE_SIGNAL_H
 #if defined(__aarch64__)
 #include <aarch64/signal.h>
+#elif defined(__x86_64__)
+#include <x86_64/signal.h>
 #else
 #include <i386/signal.h>
 #endif